refactor(sftp): migrate exec-capability probe into core
Push the desktop-only exec-capability probe onto the core SSH exec
primitive it already orchestrates, part of the #2104 SFTP convergence.

The probe marker `termihub_exec_probe_ok`, the stdout/exit discriminator,
and the probe orchestration lived in `src-tauri/src/files/sftp.rs`, with
the same marker constant copied into `core/tests/ssh_exec_with_stdin.rs`
under a hand-maintained "keep in sync" comment. That triplication is
exactly the drift the convergence targets.

Move `EXEC_PROBE_MARKER`, `exec_probe_indicates_capability`, and a new
`probe_exec_capability(session)` into `core::backends::ssh::exec`, beside
`ssh_exec_with_stdin`. `SftpSession::has_exec_capability` now wraps the
core async probe in its `block_in_place` shim with byte-identical logic
(same `echo` command, same decision, transport error → not capable). The
three probe-discriminator unit tests move to core; the core SSH exec
integration tests now exercise the real `probe_exec_capability` path
instead of re-deriving the decision, and drop their private marker copy.

Behavior-preserving: no change to the probe command, the decision, or the
`sftp_has_exec_capability` command surface. Remaining desktop-only SFTP
capabilities (elevated writes, check_writable, transfer queue, realpath,
remote_size) are deferred to follow-ups.
