+++ src-tauri/src/files/sftp.rs
@@ -9,7 +9,9 @@ use tokio::io::{AsyncReadExt, AsyncWriteExt};
 use tracing::{debug, info, warn};
 
 use termihub_core::backends::ssh::handler::SshSession;
-use termihub_core::backends::ssh::{sftp, ssh_exec_with_stdin, SshExecOutput};
+use termihub_core::backends::ssh::{
+    probe_exec_capability, sftp, ssh_exec_with_stdin, SshExecOutput,
+};
 use termihub_core::files::FileEntry;
 
 use crate::terminal::backend::SshConfig;
@@ -43,24 +45,6 @@ fn drain_sessions<V>(sessions: &Mutex<HashMap<String, V>>) -> usize {
     let count = guard.len();
     guard.clear();
     count
-}
-
-/// Marker echoed by the exec-capability probe.
-///
-/// A working shell/exec channel echoes it back on stdout; an SFTP-only
-/// connection (`ForceCommand internal-sftp`) or a relayed connection cannot run
-/// the command, so the marker never reaches stdout. Kept in sync with the
-/// integration test in `core/tests/ssh_exec_with_stdin.rs`.
-const EXEC_PROBE_MARKER: &str = "termihub_exec_probe_ok";
-
-/// Decide, from an exec probe's captured output, whether the connection can run
-/// remote commands (i.e. an exec channel is usable).
-///
-/// Keyed off the marker in stdout (not just a zero exit) so an SFTP-only server
-/// that quietly closes the forced-subsystem channel with exit 0 is still
-/// reported as not capable.
-fn exec_probe_indicates_capability(output: &SshExecOutput) -> bool {
-    output.exit_status == 0 && output.stdout.contains(EXEC_PROBE_MARKER)
 }
 
 /// Authoritative writability of a specific remote file, as decided by an SFTP
@@ -234,14 +218,9 @@ impl SftpSession {
     /// know whether privilege-elevated writes — which need a shell to run
     /// `sudo` — are possible for this connection.
     pub fn has_exec_capability(&self) -> bool {
-        let probe = format!("echo {EXEC_PROBE_MARKER}");
         tokio::task::block_in_place(|| {
-            tokio::runtime::Handle::current().block_on(async {
-                match ssh_exec_with_stdin(&self.session, &probe, "").await {
-                    Ok(output) => exec_probe_indicates_capability(&output),
-                    Err(_) => false,
-                }
-            })
+            tokio::runtime::Handle::current()
+                .block_on(async { probe_exec_capability(&self.session).await })
         })
     }
 
@@ -661,36 +640,6 @@ impl SftpManager {
 mod tests {
     use super::*;
     use std::thread;
-
-    fn probe_output(stdout: &str, exit_status: i32) -> SshExecOutput {
-        SshExecOutput {
-            stdout: stdout.to_string(),
-            exit_status,
-            ..Default::default()
-        }
-    }
-
-    /// A shell connection echoes the probe marker with exit 0 → capable.
-    #[test]
-    fn exec_probe_true_when_marker_echoed_and_exit_zero() {
-        let output = probe_output(&format!("{EXEC_PROBE_MARKER}\n"), 0);
-        assert!(exec_probe_indicates_capability(&output));
-    }
-
-    /// An SFTP-only channel yields no marker (e.g. binary SFTP bytes, even at
-    /// exit 0) → not capable.
-    #[test]
-    fn exec_probe_false_when_marker_absent() {
-        let output = probe_output("\u{0}\u{0}sftp-subsystem-bytes", 0);
-        assert!(!exec_probe_indicates_capability(&output));
-    }
-
-    /// A non-zero exit means the probe did not run cleanly → not capable.
-    #[test]
-    fn exec_probe_false_on_nonzero_exit() {
-        let output = probe_output(&format!("{EXEC_PROBE_MARKER}\n"), 1);
-        assert!(!exec_probe_indicates_capability(&output));
-    }
 
     /// Build a synthetic SFTP status error for the given status code.
     fn status_error(code: StatusCode) -> RusshSftpError {
